+++ CHANGELOG.md
@@ -20,7 +20,7 @@
   - Add support for the tags of Virtual Cache for Redis (#263)
   - Add a new configuration `kafka_namespace` to prefix the kafka topic names (#277)
   - Add log reporter support for loguru (#276)
-  - Add **experimental** support for explicit os.fork(), restarts agent in new process for Python 3.8+ (#286)
+  - Add **experimental** support for explicit os.fork(), restarts agent in forked process (#286)
   - Add **experimental** sw-python CLI flag -p/--prefork to enable non-intrusive uWSGI and Gunicorn postfork support (#288)
 
 - Plugins: